Conflicts:
	src/main/java/egovframework/ezEKP/ezApprovalG/service/EzApprovalGService.java
	src/main/java/egovframework/ezEKP/ezApprovalG/service/impl/EzApprovalGServiceImpl.java
	src/main/java/egovframework/ezEKP/ezApprovalG/web/EzApprovalGController.java
	src/main/java/egovframework/ezEKP/ezPortal/web/EzPortalController.java
	src/main/webapp/WEB-INF/jsp/ezPortal/portalWpNewApprMail.jsp
        src/main/webapp/WEB-INF/jsp/ezPortal/portalWpTotalSection.jsp

+++ src/main/webapp/WEB-INF/jsp/ezEmail/mailLeft.jsp
@@ -860,13 +860,17 @@
 			      	//지운편지함의 메일 영구삭제
 			        if (folderPath == trashBoxURL) {
 			            if (confirm("<spring:message code='ezEmail.t470' />")) {
-			                delete_mail(folderPath, true, "");
+				            if (confirm("<spring:message code='ezEmail.ksa03' />")) {
+				                delete_mail(folderPath, true, "");
+				            }
 			            }
 			        }
-			      	//편지함의 메일 지운편지함으로 이동
+			      	//편지함의 메일 지운편지함으로 이동  
 			        else {
 			            if (confirm("<spring:message code='ezEmail.t475' />")) {
-			                delete_mail(folderPath, false, trashBoxURL);
+			            	if (confirm("<spring:message code='ezEmail.ksa04' />")) {
+				                delete_mail(folderPath, false, trashBoxURL);	
+			            	}
 			            }
 			        }
 			      	
@@ -1075,7 +1079,7 @@
 	        </ul>
 	        <h2><span onclick="Address_Menu_Click();" style="width: 100%; display: inline-block;"><spring:message code="ezEmail.t99000041" /></span></h2>
 	        <ul>
-	            <div class="tree" style="height: 100%; background-color: #ffffff; border-bottom: 1px solid #eaeaea; overflow: auto; padding-left: 20px;" id="AddressTreeView"></div>
+	            <div class="tree" style="height: 100%; background-color: #ffffff; border-bottom: 1px solid #eaeaea; overflow-x:hidden; overflow-y: auto; padding-left: 20px;" id="AddressTreeView"></div>
 	            <li><span id='Address_Search' onclick="address_Search();" style="width: 100%; display: inline-block;"><spring:message code="ezEmail.t99000042" /></span></li>
 	            <li style="border-bottom-color:#e8e8e8" evt="0"><span onclick="address_foldermanage()" style="width: 100%; display: inline-block;"><spring:message code="ezEmail.t99000043" /></span></li>
 	        </ul>
